Move all task components inside a folder task. More cleaning on tasks...
